improve folder checks (with new subfolders)

+++ main.py
@@ -138,12 +138,14 @@ class DevImageEditApp(App):
         self.check_folder(".temp")
         self.check_folder(ImageWorkDirManager.work_image_path)
         self.check_folder(Engine.work_dir)
+        self.check_folder(os.path.join(Engine.work_dir, "remove_bg_work_dir"))
         SettingsManager().check_settings_file()
 
     def on_stop(self):
         self.check_folder(".temp")
         self.check_folder(ImageWorkDirManager.work_image_path)
         self.check_folder(Engine.work_dir)
+        self.check_folder(os.path.join(Engine.work_dir, "remove_bg_work_dir"))
 
     def build(self):
         self.icon = "images/logo_dev_icon_editor.jpg"
@@ -161,7 +163,7 @@ class DevImageEditApp(App):
                 file_path = os.path.join(folder_path, filename)
                 if os.path.isfile(file_path):
                     os.remove(file_path)
-                elif os.path.isdir(file_path):
+                elif os.path.isdir(file_path) and file_path != os.path.join(Engine.work_dir, "remove_bg_work_dir"):
                     os.rmdir(file_path)
         else:
             os.mkdir(folder_path)
